+++ packages/redis/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @idearium/redis
 
+## Unreleased
+
+### Changed
+
+-   The new default is to always return a new redis connection.
+-   You can now pass in `reuse` and `reuseIndex` to have an existing connection returned.
+
 ## v1.0.0-beta.1
 
 -   First version of the package.